java-practice/test-driven-development/src/test/java/chap07/step1/AutoDebitRegisterTest.java
package chap07.step1;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class AutoDebitRegisterTest {
    private AutoDebitRegister register;

    @BeforeEach
    void setUp() {
        CardNumberValidator validator = new CardNumberValidator();
        JpaAutoDebitInfoRepository repository = new JpaAutoDebitInfoRepository();
        register = new AutoDebitRegister(validator, repository);
    }

    @Test
    void validCard() {
        // 업체에서 받은 테스트용 유효한 카드번호 사용
        final AutoDebitReq req = new AutoDebitReq("user1", "1234123412341234");
        final RegisterResult result = this.register.register(req);

        assertThat(result.getValidity()).isEqualTo(CardValidity.VALID);
    }

    @Test
    void theftCard() {
        // 업체에서 받은 도난 테스트용 카드번호 사용
        final AutoDebitReq req = new AutoDebitReq("user1", "1234567890123456");
        RegisterResult result = this.register.register(req);

        assertThat(result.getValidity()).isEqualTo(CardValidity.THEFT);
    }
}
